Decompile NP3 EntityHammer
