+++ config/splat.us.stnp3.yaml
@@ -124,10 +124,10 @@ segments:
       - [0x320B4, .rodata, 49F98] # EntityBloodyZombie data
       - [0x320DC, rodata]
       - [0x321EC, .rodata, 4B018]
-      - [0x3224C, rodata]
+      - [0x3224C, .rodata, 4E69C] #EntityHammer
       - [0x322CC, rodata] # EntityGurkhaBodyParts
       - [0x322F8, rodata] # EntityGurkha
-      - [0x32380, .rodata, 4E69C] # EntityGurkhaSword
+      - [0x32380, .rodata, 4F254] # EntityGurkhaSword
       - [0x323E4, rodata]
       - [0x32464, .rodata, prim_helpers]
       - [0x3246C, c]
@@ -153,6 +153,7 @@ segments:
       - [0x49F98, c] # EntityBloodyZombie
       - [0x4B018, c]
       - [0x4E69C, c]
+      - [0x4F254, c]
       - [0x51F38, c, prim_helpers]
       - [0x52768, sbss]
   - [0x53434]
