+++ config/symbols.us.stnz0.txt
@@ -1,6 +1,4 @@
-g_SpriteBanks = 0x8018002C;
 g_Cluts = 0x80180160;
-g_TileLayers = 0x80180394;
 g_EntityGfxs = 0x80180888;
 g_pStObjLayoutHorizontal = 0x801808EC;
 g_pStObjLayoutVertical = 0x801809C0;
@@ -53,12 +51,7 @@ EntityLargeGaibonProjectile = 0x801B6BBC;
 EntityElevator = 0x801B7034;
 EntityMariaCutscene = 0x801B7D58;
 EntityMaria = 0x801B8F94;
-Update = 0x801B9504;
-UpdateStageEntities = 0x801B9800;
-HitDetection = 0x801B9908;
 EntityDamageDisplay = 0x801BAA20;
-InitRoomEntities = 0x801BB920;
-UpdateRoomPosition = 0x801BBA98;
 EntityRedDoor = 0x801BBCB4;
 EntityDummy = 0x801BD6F8;
 EntityPrizeDrop = 0x801BE30C;
