+++ config/symbols.us.txt
@@ -847,6 +847,8 @@ g_LoadFile = 0x8006BAFC;
 g_CastleMap = 0x8006BB74;
 g_CurrentBuffer = 0x8006C37C;
 _snd_ev_flag = 0x8006C380;
+D_8006C384 = 0x8006C384;
+D_8006C38C = 0x8006C38C;
 kMaxPrograms = 0x8006C394;
 g_CdStep = 0x8006C398;
 g_backbufferX = 0x8006C39C;
