+++ config/undefined_syms.hd.txt
@@ -2,7 +2,8 @@ g_CastleFlags = 0x8003BDEC;
 D_8003C908    = 0x8003C908;
 g_GpuBuffers  = 0x8003CB08;
 g_api         = 0x8003C774;
-
+D_8006C384    = 0x8006C384;
+D_8006C38C    = 0x8006C38C;
 g_Clut        = 0x8006CBCC;
 
 g_Player      = 0x80072BD0;
